requirements.txt
altgraph==0.17
asgiref==3.2.3
attrs==19.3.0
autoit==0.1
Automat==0.8.0
backcall==0.1.0
beautifulsoup4==4.8.2
beautifulsoup442==4.8.2
bleach==3.1.5
bs4==0.0.1
certifi==2020.4.5.1
cffi==1.13.2
chardet==3.0.4
chromedriver-autoinstaller==0.2.0
Click==7.0
colorama==0.4.3
configparser==5.0.0
constantly==15.1.0
crayons==0.3.0
cryptography==2.8
cssselect==1.1.0
cycler==0.10.0
Cython==0.29.22
decorator==4.4.2
defusedxml==0.6.0
Django==3.0.2
django-appconf==1.0.3
django-bootstrap4==1.1.1
django-crispy-forms==1.9.0
django-imagekit==4.0.2
django-jquery==3.1.0
entrypoints==0.3
Flask==1.1.1
future==0.18.2
geographiclib==1.50
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
ipykernel==5.2.1
ipython==7.14.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
jedi==0.17.0
Jinja2==2.10.3
jsonschema==3.2.0
jupyter @ file:///C:/jupyter-1.0.0-py2.py3-none-any.whl
jupyter-client==6.1.3
jupyter-console==6.1.0
jupyter-core==4.6.3
kiwisolver==1.1.0
lxml==4.5.0
MarkupSafe==1.1.1
matplotlib==3.1.3
mistune==0.8.4
mysql-connector-python==8.0.20
mysqlclient==1.4.6
nbconvert==5.6.1
nbformat==5.0.6
notebook==6.0.3
numpy==1.18.1
packaging==20.3
pandas==1.0.1
pandocfilters==1.4.2
parsel==1.5.2
parso==0.7.0
pefile==2019.4.18
pickleshare==0.7.5
pilkit==2.0
Pillow==7.1.2
prometheus-client==0.7.1
prompt-toolkit==3.0.5
Protego==0.1.16
protobuf==3.11.3
proxyscrape==0.3.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.19
PyDispatcher==2.0.5
Pygments==2.6.1
PyHamcrest==2.0.0
PyInstaller==3.6
pyOpenSSL==19.1.0
pyparsing==2.4.6
pyrsistent==0.16.0
PySide2==5.14.2.1
python-dateutil==2.8.1
python-slugify==4.0.1
pytz==2019.3
pywin32==227
pywin32-ctypes==0.2.0
pywinpty==0.5.7
pyzmq==19.0.1
qtconsole==4.7.3
QtPy==1.9.0
queuelib==1.5.0
requests==2.23.0
reverse-geocoder==1.5.1
scipy==1.6.0
Scrapy @ file:///C:/Scrapy-2.1.0-py3-none-any.whl
scrapy-proxy-pool==0.1.9
scrapy-user-agents==0.1.1
selenium==3.141.0
Send2Trash==1.5.0
service-identity==18.1.0
shiboken2==5.14.2.1
six==1.14.0
soupsieve==1.9.5
spark-parser==1.8.9
sqlparse==0.3.0
terminado==0.8.3
testpath==0.4.4
text-unidecode==1.3
tornado==6.0.4
traitlets==4.3.3
Twisted==19.10.0
ua-parser==0.10.0
uncompyle6==3.7.4
urllib3==1.25.9
urlopen==1.0.0
user-agents==2.2.0
w3lib==1.21.0
wcwidth==0.1.9
webdriver-manager==2.5.3
webencodings==0.5.1
Werkzeug==0.16.0
widgetsnbextension==3.5.1
xdis==5.0.8
zope.interface==4.7.1
